Initial guard funcionality.
This is intended to reduce the boilerplate code that the project will
introduce in future.

It is simple class that will be used for adding so called guards (name
borrowed from Haskell).
Right now only null reference guard is implemented, but many may follow.

Guard functionality is placed in new project (Optional.Internals), but
should not be intended for packaging hence the name "Internals".